refactor readers

+++ include/Readers/ICursorReader.hpp
@@ -5,25 +5,24 @@
 
 namespace RedatamLib {
 
-class ICursorReader
-{
+class ICursorReader {
 public:
-    virtual ~ICursorReader() = default;
+  virtual ~ICursorReader() = default;
 
-    virtual void Close() = 0;
-    virtual bool IsLastPos() const = 0;
-    virtual void Open() = 0;
-    virtual double ReadDouble() = 0;
-    virtual uint32_t ReadInt16() = 0;
-    virtual uint32_t ReadInt32() = 0;
-    virtual uint32_t ReadLastInt32() = 0;
-    virtual uint64_t ReadNumber() = 0;
-    virtual std::string ReadString() = 0;
+  virtual void Close() = 0;
+  virtual bool IsLastPos() const = 0;
+  virtual void Open() = 0;
+  virtual double ReadDouble() = 0;
+  virtual uint16_t ReadInt16() = 0;
+  virtual uint32_t ReadInt32() = 0;
+  virtual uint32_t ReadLastInt32() = 0;
+  virtual uint64_t ReadNumber() = 0;
+  virtual std::string ReadString() = 0;
 
-    virtual int64_t Length() const = 0;
-    virtual uint32_t ReadInt32At(int64_t pos) = 0;
+  virtual int64_t Length() const = 0;
+  virtual uint32_t ReadInt32At(int64_t pos) = 0;
 };
 
-}
+} // namespace RedatamLib
 
 #endif // ICURSORREADER_HPP
